refactor: changed parser location
