+++ src/main/java/ui/CalculatorController.java
@@ -1,6 +1,7 @@
 package ui;
 
 import calculator.*;
+import calculator.Parser;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
